k-diffusion/requirements.txt
accelerate
clean-fid
einops
jsonmerge
kornia
Pillow
resize-right
scikit-image
scipy
torch
torchdiffeq
torchvision
tqdm
wandb
git+https://github.com/openai/CLIP
